Update dependencies, workflows, and increase Java and Scala versions
Updated the dependencies in the build.sbt, as well as increased the Scala version in use. Updated the Java version in the Github workflows, and updated the versions of the Github actions used, as well as added a new workflow for dependency graph. Lastly, added a dependabot configuration to automatically keep the Github Actions up to date.